resume-analyzer/requirements.txt
flask==3.0.3
pdfplumber==0.11.0
python-docx==1.1.2
google-generativeai
gunicorn==22.0.0